refactor(icon-picker): 重組測試展示檔案到 demo 目錄
- 移動 IconPickerDevTool.vue 到 demo/ 目錄
- 移動 IconPickerProxy.vue 到 demo/ 目錄
- 更新所有相關的導入路徑
- 保持測試檔案和模組匯出的正確性

目的：
- 將測試展示用途的檔案與核心元件分離
- 提升專案結構的清晰度和可維護性
- 便於後續開發和測試管理
