+++ resources/js/components/test/IconPickerTestPage.vue
@@ -130,8 +130,8 @@
 
 <script>
 import { ref, computed } from 'vue'
-import IconPickerProxy from '@/features/icon-picker/components/IconPickerProxy.vue'
-import IconPickerDevTool from '@/features/icon-picker/components/IconPickerDevTool.vue'
+import IconPickerProxy from '@/features/icon-picker/demo/IconPickerProxy.vue'
+import IconPickerDevTool from '@/features/icon-picker/demo/IconPickerDevTool.vue'
 import IconDisplay from '@/components/common/IconDisplay.vue'
 
 export default {
