+++ resources/js/features/icon-picker/demo/IconPickerProxy.vue
